Requesting data from mirrors where we run tests
Given that we need to allow developers to run tests, we created these 2
mirrors to run pipelines on the default branch for developers. This way,
we don't need to let developers run pipelines on the default branch for
the actual projects.

We need to have data for these mirrors so that we can analyse how we run
the test pipelines.